refactor: relocate gene-reference, panel, and HPO types to src/shared/
Move pure data type interfaces out of main-process files into new shared
type modules (gene-reference.ts, panels.ts) and inline HpoTerm in
api-enrichment.ts, eliminating all remaining shared→main import violations.
