+++ src/main/database/PanelRepository.ts
@@ -4,52 +4,8 @@ import { sqlPlaceholders } from './sql-utils'
 
 // ── Types ────────────────────────────────────────────────────
 
-export interface GenomicInterval {
-  chr: string // chromosome name matching variant data format
-  start: number // 1-based, with padding applied
-  end: number // 1-based, with padding applied
-}
-
-export interface CreatePanelInput {
-  name: string
-  description?: string | null
-  version?: string | null
-  source: string
-  sourceId?: string | null
-  sourceMetadata?: Record<string, unknown> | null
-}
-
-export interface PanelRow {
-  id: number
-  name: string
-  description: string | null
-  version: string | null
-  source: string
-  source_id: string | null
-  source_metadata: string | null
-  created_at: number
-  updated_at: number
-}
-
-export interface PanelWithCount extends PanelRow {
-  gene_count: number
-}
-
-export interface PanelGeneRow {
-  id: number
-  panel_id: number
-  hgnc_id: string
-  symbol: string
-}
-
-export interface ActivePanelRow {
-  case_id: number
-  panel_id: number
-  padding_bp: number
-  activated_at: number
-  panel_name: string
-  gene_count: number
-}
+import type { GenomicInterval, CreatePanelInput, PanelRow, PanelWithCount, PanelGeneRow, ActivePanelRow } from '../../shared/types/panels'
+export type { GenomicInterval, CreatePanelInput, PanelRow, PanelWithCount, PanelGeneRow, ActivePanelRow }
 
 // ── Repository ───────────────────────────────────────────────
