+++ src/shared/types/api.ts
@@ -84,19 +84,8 @@ import type { FilterPreset, FilterPresetCreate, FilterPresetUpdate } from './fil
 import type { LogMessage } from './log'
 import type { TranscriptAnnotation, TranscriptInsertRow } from './transcript'
 import type { DatabaseOverview } from './database-overview'
-import type {
-  GeneValidationResult,
-  GeneAutocompleteResult,
-  GeneRefInfo,
-  AssemblyInfo
-} from '../../main/database/GeneReferenceDb'
-import type {
-  PanelRow,
-  PanelWithCount,
-  PanelGeneRow,
-  ActivePanelRow
-} from '../../main/database/PanelRepository'
-import type { PanelAppSearchResult } from '../../main/services/api/PanelAppClient'
+import type { GeneValidationResult, GeneAutocompleteResult, GeneRefInfo, AssemblyInfo } from './gene-reference'
+import type { PanelRow, PanelWithCount, PanelGeneRow, ActivePanelRow, PanelAppSearchResult } from './panels'
 import type {
   ProteinMappingResult,
   ProteinDomainResult,
